+++ src/test/java/com/bitbank/test/services/v1/BankAccountTransactionsServiceTest.java
@@ -1,39 +1,42 @@
-package com.bitbank.test.repository.v1;
+package com.bitbank.test.services.v1;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.bitbank.entities.v1.BankAccountTransactionsEntity;
 import com.bitbank.repository.v1.BankAccountTransactionsRepository;
+import com.bitbank.services.v1.BankAccountTransactionsService;
 
-@ExtendWith(SpringExtension.class)
-@DataJpaTest
-class BankAccountTransactionsRepositoryTest {
-	
-	@Autowired
+@ExtendWith(MockitoExtension.class)
+class BankAccountTransactionsServiceTest {
+	@Mock
 	private BankAccountTransactionsRepository bankAccountTransactionsRepository;
 	
-	/**
-	 * Index
-	 */
+	@InjectMocks
+	private BankAccountTransactionsService bankAccountTransactionsService;
+	
 	@Test
-	void testCanFindAll() {
+	void get_testCanIndex() {
+		
 		var transaction01 = BankAccountTransactionsEntity
 				.builder()
+				.id(1L)
 				.amount(new BigDecimal(99.99))
 				.purpose("First transaction")
 				.build();
 		
 		var transaction02 = BankAccountTransactionsEntity
 				.builder()
+				.id(2L)
 				.amount(new BigDecimal(150))
 				.purpose("Second transaction")
 				.build();
@@ -45,12 +48,9 @@ class BankAccountTransactionsRepositoryTest {
 		result.add(transaction01);
 		result.add(transaction02);
 		
-		assertEquals(result, bankAccountTransactionsRepository.findAll());
-	}
-	
-	@Test
-	void testCanFindAllWithoutItems() {
-		var result = new ArrayList<BankAccountTransactionsEntity>();
-		assertEquals(result, bankAccountTransactionsRepository.findAll());
+		when(bankAccountTransactionsRepository.findAll()).thenReturn(result);
+		
+		assertEquals(result, bankAccountTransactionsService.index());
+		
 	}
 }