+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Add CI/CD
 + Add Bank Account Transactions entity
 + Add Bank Account Transactions repository
++ Add Bank Account Transactions service
 
 ## 0.1.0
 + First commit